pyproject.toml
[project]
name = "gas-sozai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
green-brown = "python src/green_brown.py"
